pyproject.toml
[project]
name = "global-technologies"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.9.0",
    "django-json-widget>=2.1.1",
    "django-ninja>=1.5.1",
    "django-ninja-jwt[crypto]>=5.4.2",
    "django-q2>=1.9.0",
    "django-redis>=6.0.0",
    "django-unfold>=0.74.1",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.1.1",
    "razorpay>=2.0.0",
]
